play fix

+++ python_xiami_api/submitter.py
@@ -13,7 +13,7 @@ def refresh_token():
 	url
 def upload(path,fname):
 	import requests as r
-	url_file=  "https://pcs.baidu.com/rest/2.0/pcs/file"
+	url_file=  "https://c.pcs.baidu.com/rest/2.0/pcs/file"
 	token="3.9f695526212e72665604185c73ae9a66.2592000.1385028974.4093755095-1182756"
 	print "@uploading",repr(path)
 	res=r.post(url_file,
@@ -24,25 +24,28 @@ def upload(path,fname):
 			"ondup":"overwrite",
 		},
 		files={
-			"file":file(fname,'rb')
-		}
+			"file":file(fname,'rb').read()
+		},
+		verify=False
 	)
-	print res.content
+	print res.content,res.status_code
+	return res.status_code
 def push():
-	#write version info, ID is unix time
+	# write version info, ID is unix time
 	file(f_version,'w').write(str(time.time()))
 	z=zipfile.ZipFile(f_zip,'w')
 	def allows(x):
 		if x=="access_token" or x=="latest":
 			return False
 		ext=os.path.splitext(x.lower())[1]
-		return ext not in ['.txt','.pdb','.log','.zip']
+		return ext not in ['.txt','.pdb','.log','.zip','.suo','.dat']
 	for f in filter(allows,os.listdir(f_xiami)):
 		z.write(os.path.join(f_xiami,f),f)
 	z.close()
 	path="/apps/folder1/"
-	upload(path+"version.txt",f_version)
-	upload(path+"latest.zip",f_zip)
-	os.remove(f_version)
-	os.remove(f_zip)
+	status=upload(path+"version.txt",f_version)
+	status=upload(path+"latest.zip",f_zip)
+	if status==200:
+		os.remove(f_version)
+		os.remove(f_zip)
 push()
